requirements.txt
annotated-types==0.7.0
anyio==4.12.1
av==16.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.15
ctranslate2==4.6.3
faiss-cpu==1.13.2
faster-whisper==1.2.1
filelock==3.20.3
flatbuffers==25.12.19
fsspec==2026.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
ollama==0.6.1
onnxruntime==1.23.2
packaging==26.0
protobuf==6.33.4
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.6.0
pypiwin32==223
pyreadline3==3.5.4
pyttsx3==2.99
pywin32==311
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.0
setuptools==80.10.1
sounddevice==0.5.4
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.1
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
